+++ POO/src/Balanceo.java
@@ -1,4 +1,4 @@
-package src;
+
 
 import java.awt.EventQueue;
