+++ POO/src/Estequio.java
@@ -1,4 +1,4 @@
-package src;
+
 
 import java.awt.EventQueue;
 
@@ -6,18 +6,18 @@ import javax.swing.JFrame;
 import javax.swing.JPanel;
 import javax.swing.border.EmptyBorder;
 
-public class Nomenclatura extends JFrame {
+public class Estequio extends JFrame {
 
 	private static final long serialVersionUID = 1L;
 	private JPanel contentPane;
 
 	/**
-
+	 * Launch the application.
 	public static void main(String[] args) {
 		EventQueue.invokeLater(new Runnable() {
 			public void run() {
 				try {
-					Nomenclatura frame = new Nomenclatura();
+					Estequio frame = new Estequio();
 					frame.setVisible(true);
 				} catch (Exception e) {
 					e.printStackTrace();
@@ -26,9 +26,10 @@ public class Nomenclatura extends JFrame {
 		});
 	}
 
-
+	/**
+	 * Create the frame.
 	 */
-	public Nomenclatura() {
+	public Estequio() {
 		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		setBounds(100, 100, 450, 300);
 		contentPane = new JPanel();
