+++ POO/src/MasaMolar.java
@@ -1,0 +1,18 @@
+
+
+import java.awt.BorderLayout;
+import java.awt.EventQueue;
+import java.awt.GridLayout;
+import java.awt.Image;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import javax.swing.*;
+import javax.swing.border.EmptyBorder;
+
+public class MasaMolar extends JFrame {
+    private JPanel contentPane;
+    private JTextPane instruccion;
+    private JLabel cartelavatar;
+    private JLabel infouser;
+    private JButton botonAvatar;
+}
