+++ POO/src/Perfil.java
@@ -1,4 +1,4 @@
-package src;
+
 
 import java.awt.BorderLayout;
 import java.awt.EventQueue;
