--- POO/src/module-info.java
@@ -1,9 +1,0 @@
-/**
- * 
- */
-/**
- * 
- */
-module POO {
-	requires java.desktop;
-}
